+++ services/horizon/internal/ingest/main_test.go
@@ -17,7 +17,6 @@ import (
 	"github.com/stellar/go/ingest/ledgerbackend"
 	"github.com/stellar/go/services/horizon/internal/db2/history"
 	"github.com/stellar/go/services/horizon/internal/ingest/processors"
-	"github.com/stellar/go/services/horizon/internal/verify"
 	"github.com/stellar/go/support/db"
 	"github.com/stellar/go/support/errors"
 	logpkg "github.com/stellar/go/support/log"
@@ -76,7 +75,7 @@ func TestCheckVerifyStateVersion(t *testing.T) {
 	assert.Equal(
 		t,
 		CurrentVersion,
-		verify.StateVerifierExpectedIngestionVersion,
+		StateVerifierExpectedIngestionVersion,
 		"State verifier is outdated, update it, then update StateVerifierExpectedIngestionVersion value",
 	)
 }