+++ services/horizon/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this
 file. This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
-- moved ingest/verify to be internally located at services/horizon/internal/verify package. [5670](https://github.com/stellar/go/issues/5670)
+- moved ingest/verify to be internally located at services/horizon/internal/ingest package. [5670](https://github.com/stellar/go/issues/5670)
 
 ## 23.0.0